ai-ml/use-cases/02_eligibility_scoring_360_profile/requirements.txt
# Eligibility Scoring & 360° Profiles Use Case
# Python dependencies

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ML Models
xgboost>=2.0.0

# MLflow
mlflow>=2.8.0

# Graph analysis
networkx>=3.1  # Fallback option (not used if Neo4j enabled)
scipy>=1.10.0

# Neo4j driver (for graph database)
neo4j>=5.0.0

# Database
psycopg2-binary>=2.9.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
rapidfuzz>=3.0.0
phonetics>=1.0.0
geopy>=2.3.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (optional)
jupyterlab>=4.0.0
ipykernel>=6.25.0
